Some more unit tests
Also removed a bunch of println statements

+++ CommonTower/src/main/java/com/kag/commontower/entities/parts/WeaponPart.java
@@ -43,7 +43,7 @@ public class WeaponPart implements IPart {
 		return attackSpeed;
 	}
 
-	public void setAttackSpeed(int attackSpeed) {
+	public void setAttackSpeed(float attackSpeed) {
 		this.attackSpeed = attackSpeed;
 	}
 }
